layout like a parent component to main and favourites page. Favourites page working by Outlet. All the logic of Main page now in Layout
